extract segment update to a seperate file
