+++ examples/oncall_routing/greedy.py
@@ -7,24 +7,10 @@ from maro.simulator import Env
 from maro.simulator.scenarios.oncall_routing.common import Action, OncallRoutingPayload
 from maro.utils import set_seeds
 
+from examples.oncall_routing.utils import refresh_segment_index
+
 set_seeds(0)
 
-
-def _is_equal_segment(action1: Action, action2: Action) -> bool:
-    return (action1.route_name, action1.insert_index) == (action2.route_name, action2.insert_index)
-
-def _refresh_segment_index(actions: List[Action]) -> List[Action]:
-    # Add segment index if multiple orders are sharing a same insert index.
-    actions.sort(key=lambda action: (action.route_name, action.insert_index))
-    segment_index = 0
-    for idx in range(len(actions) - 1):
-        if _is_equal_segment(actions[idx], actions[idx + 1]):
-            segment_index += 1
-            actions[idx + 1].in_segment_order = segment_index
-        else:
-            segment_index = 0
-
-    return actions
 
 def _get_actions(running_env: Env, event: OncallRoutingPayload) -> List[Action]:
     tick = running_env.tick
@@ -54,7 +40,7 @@ def _get_actions(running_env: Env, event: OncallRoutingPayload) -> List[Action]:
         if chosen_route_name is not None:
             actions.append(Action(order_id=oncall_order.id, route_name=chosen_route_name, insert_index=insert_idx))
 
-    actions = _refresh_segment_index(actions)
+    actions = refresh_segment_index(actions)
 
     return actions
